Feat: edit Home page to redirect to Login if not logged in...
- add basic styles to Navbar and Header
