+++ src/assets/styles/Header.scss
@@ -1,0 +1,7 @@
+.Header {
+  background-color: #41b5e8;
+  height: 50px;
+  display: flex;
+  align-items: center;
+  padding: 10px;
+}
