+++ src/components/Header.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { IoIosArrowBack } from 'react-icons/io';
+import '../assets/styles/Header.scss';
 
 const Header = () => (
   <div className="Header"><IoIosArrowBack /></div>
